+++ classes/dataflow/decode/yaml.php
@@ -1,0 +1,11 @@
+<?php defined('SYSPATH') or die('No direct script access.');
+/**
+ * Dataflow Decode YAML Library
+ * 
+ * @package		Dataflow
+ * @category	Base
+ * @author		Micheal Morgan <micheal@morgan.ly>
+ * @copyright	(c) 2011 Micheal Morgan
+ * @license		MIT
+ */
+class Dataflow_Decode_Yaml extends Kohana_Dataflow_Decode_Yaml {}