+++ classes/dataflow/encode.php
@@ -1,0 +1,11 @@
+<?php defined('SYSPATH') or die('No direct script access.');
+/**
+ * Dataflow Encode Library
+ * 
+ * @package		Dataflow
+ * @category	Base
+ * @author		Micheal Morgan <micheal@morgan.ly>
+ * @copyright	(c) 2011 Micheal Morgan
+ * @license		MIT
+ */
+abstract class Dataflow_Encode extends Kohana_Dataflow_Encode {}