+++ classes/dataflow/encode/php.php
@@ -1,0 +1,11 @@
+<?php defined('SYSPATH') or die('No direct script access.');
+/**
+ * Dataflow Encode PHP Library
+ * 
+ * @package		Dataflow
+ * @category	Base
+ * @author		Micheal Morgan <micheal@morgan.ly>
+ * @copyright	(c) 2011 Micheal Morgan
+ * @license		MIT
+ */
+class Dataflow_Encode_Php extends Kohana_Dataflow_Encode_Php {}