+++ classes/kohana/dataflow/encode/json.php
@@ -1,0 +1,38 @@
+<?php defined('SYSPATH') or die('No direct script access.');
+/**
+ * Dataflow Encode JSON Driver
+ * 
+ * @package		Dataflow
+ * @category	Base
+ * @author		Micheal Morgan <micheal@morgan.ly>
+ * @copyright	(c) 2011 Micheal Morgan
+ * @license		MIT
+ */
+class Kohana_Dataflow_Encode_Json extends Dataflow_Encode
+{	
+	/**
+	 * Get content type
+	 * 
+	 * @access	public
+	 * @return	string
+	 */
+	public function get_content_type()
+	{
+		return 'application/json';
+	}	
+	
+	/**
+	 * Encode
+	 * 
+	 * @access	protected
+	 * @return	string
+	 */
+	protected function _encode(array $data)
+	{
+		// Force return of object if array empty
+		if (empty($data))
+			return '{}';
+		else
+			return json_encode($data);
+	}	
+}