+++ config/dataflow.php
@@ -1,0 +1,11 @@
+<?php defined('SYSPATH') or die('No direct script access.');
+return array
+(
+	'json'	=> array(),
+	'php'	=> array(),
+	'xml'	=> array
+	(
+		'attribute'	=> ':attributes'
+	),
+	'yaml'	=> array()
+);