+++ tests/kohana/DataflowTest.php
@@ -9,57 +9,111 @@
  * @copyright	(c) 2011 Micheal Morgan
  * @license		MIT
  */
-class Kohana_DataflowTest extends Unittest_TestCase
+abstract class Kohana_DataflowTest extends Unittest_TestCase
 {	
+    /**
+     * Factory to return Dataflow object configured for driver.
+     * 
+     * @access	public
+     * @return	Dataflow
+     */
+    abstract public function factory();
+    
 	/**
-	 * Data provider for JSON samples
-	 *
+	 * Sample provider
+	 * 
 	 * @access	public
 	 * @return	array
 	 */
-	public static function provider_json()
-	{
-		return array
-		(
-			// empty
-			array
-			(
-				array(),
-				'{}'
-			),
-			// indexed array
-			array
-			(
-				array('a', 'b', 'c', 'd'),
-				'["a","b","c","d"]'
-			),
-			// key/value array
-			array
-			(
-				array('key' => 'value'),
-				'{"key":"value"}'
-			)
-		);
-	}
-	
-	/**
-	 * Run input/output on Json provider
-	 * 
-	 * @covers			Dataflow::set
-	 * @covers			Dataflow::get
-	 * @covers			Dataflow::__toString
-	 * @dataProvider	provider_json
-	 */
-	public function test_json($set, $expected)
-	{
-		$dataflow = Dataflow::factory(array('output' => array('driver' => 'json')));
-		
-		$dataflow->set($set);
-	
-		$this->assertSame
-		(
-			$expected,
-			(string) $dataflow
-		);		
-	}
+    public function provider_samples()
+    {
+    	return array
+    	(
+    		array
+    		(
+    			// test empty array
+    			array()
+    		),
+    		array
+    		(
+    			// test indexed array with a variety of variable types
+    			array
+    			(
+    				'samples' => array
+    				(
+    					'sample' => array('string', 1)
+    				)
+    			)
+    		),
+    		array
+    		(
+    			// test associative array
+    			array
+    			(
+	    			'key1' => array
+	    			(
+	    				'key2_1'	=> 'value2_1', 
+	    				'key2_2'	=> array
+	    				(
+	    					'key3_1'	=> 'value_3_1',
+	    					'key3_2'	=> 'value_3_2'
+	    				)
+	    			)
+	    		)
+    		),
+    		array
+    		(
+    			// test attribute handling
+    			array
+    			(
+    				'key1' => array
+    				(
+    					'_attributes'	=> array('a_key1' => 'a_key2'),
+  						'key2'			=> array
+    					(
+    						'value2_1',
+    						'value_2_3'
+    					)
+    				)
+    			)
+    		)
+    	);
+    }
+    
+    /**
+     * Test by encoding sample and comparing original array with decoded
+     * 
+     * @covers			Dataflow::factory
+     * @covers			Dataflow::set
+     * @covers			Dataflow::get
+     * @covers			Dataflow::as_array
+     * @covers			Dataflow_Decode::factory
+     * @covers			Dataflow_Decode::get
+     * @covers			Dataflow_Decode::set
+     * @covers			Dataflow_Decode::_decode
+     * @covers			Dataflow_Encode::factory
+     * @covers			Dataflow_Encode::get
+     * @covers			Dataflow_Encode::set
+     * @covers			Dataflow_Encode::_encode
+     * @dataProvider	provider_samples
+     * @access			public
+     * @param			array
+     * @return			void
+     */
+    public function test_samples(array $sample)
+    {
+    	$dataflow = $this->factory();
+    	
+    	// Set sample input
+    	$dataflow->set($sample);
+    	
+    	// Get encoded
+    	$encoded = $dataflow->get();
+
+    	// Decode encoded
+    	$dataflow->set($encoded);
+
+    	// Test decoded matches original sample
+    	$this->assertEquals($sample, $dataflow->as_array());	
+    }
 }